app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.3

# Azure
azure-ai-formrecognizer==3.3.3
azure-ai-openai==1.0.0b1
azure-storage-blob==12.23.1
azure-identity==1.19.0

# Background jobs
celery==5.4.0
redis==5.2.0

# NLP
transformers==4.46.3
torch==2.5.1
sentencepiece==0.2.0

# Utilities
python-multipart==0.0.12
httpx==0.27.2
structlog==24.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
faker==33.1.0

# Dev
black==24.10.0
flake8==7.1.1
mypy==1.13.0
